Closes #60
Split service into three
